+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Change Log
  - Added Extension methods for Setting Collections.
  - Added Entry data structures to settings for pairing keys and values together.
  - Added extension methods to look up a collection of settings, returned as entries.
+ - Added Extension methods for Setting Collections.
+
+ ### Deprecated
+
+ - Deprecated `SettingsCollection` in favor of a generic collection of Settings.
 
 1.5.0
 -----